tsp, refactor ts code
